refactor: PublishedStatusRequest 패키지 변경
