+++ backend/src/main/java/hanglog/trip/dto/request/PublishedStatusRequest.java
@@ -1,4 +1,4 @@
-package hanglog.trip.dto;
+package hanglog.trip.dto.request;
 
 import static lombok.AccessLevel.PRIVATE;
